more removal of longs fixes for static code analysis removal of unused code in cole_encode.
git-svn-id: svn://svn.code.sf.net/p/rtf2latex2e/code@40 ec8729b6-1b83-4082-b6c8-10560a4245e1

+++ src/rtf2latex2e.h
@@ -27,7 +27,7 @@
 typedef unsigned char boolean;
 #endif
 # define	New(t)	((t *) RTFAlloc (sizeof (t)))
-# define	Swap(a, b) {double tmp;\
+# define	Swap(a, b) {int tmp;\
 						tmp = a;\
 						a = b;\
 						b = tmp;}
@@ -67,12 +67,12 @@ typedef struct
 {
 	int	count;
 	int	type;
-	long	width;
-	long	height;
-	long	goalWidth;
-	long	goalHeight;
-	long	scaleX;
-	long	scaleY;
+	int32_t	width;
+	int32_t	height;
+	int32_t	goalWidth;
+	int32_t	goalHeight;
+	int32_t	scaleX;
+	int32_t	scaleY;
 	int		llx;
 	int		lly;
 	int		urx;
